requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0

# Fine-tuning & Efficiency
peft>=0.4.0
trl>=0.4.7
bitsandbytes>=0.41.0

# Frontend & UI
gradio>=4.0.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Utilities
tqdm>=4.64.0
python-dotenv>=0.19.0
PyYAML>=6.0
plotly>=5.0.0
PyJWT>=2.6.0

# Testing
pytest>=7.0.0
pytest-mock>=3.10.0